+++ demo/scripts/controls/editor/EditorInstanceToggleablePlugins.ts
@@ -9,6 +9,7 @@ import { HyperLink } from 'roosterjs-editor-plugins/lib/HyperLink';
 import { Paste } from 'roosterjs-editor-plugins/lib/Paste';
 import { PickerPlugin } from 'roosterjs-editor-plugins/lib/Picker';
 import { TableResize } from 'roosterjs-editor-plugins/lib/TableResize';
+import { TableSelection } from 'roosterjs-editor-plugins/lib/TableSelection';
 import { Watermark } from 'roosterjs-editor-plugins/lib/Watermark';
 
 export type EditorInstanceToggleablePlugins = {
@@ -23,4 +24,5 @@ export type EditorInstanceToggleablePlugins = {
     pickerPlugin: PickerPlugin;
     contextMenu: ContextMenu<ContextMenuItem>;
     resetList: ResetListPlugin;
+    tableSelection: TableSelection;
 };
