+++ packages/roosterjs-editor-plugins/lib/TableSelection.ts
@@ -1,0 +1,1 @@
+export * from './plugins/TableSelection/index';
