+++ packages/roosterjs-editor-plugins/lib/plugins/TableSelection/index.ts
@@ -1,0 +1,1 @@
+export { default as TableSelection } from './TableSelection';
